requirements.txt
Django==3.1.7
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.2
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
